Organize nested scenes into subdirectories
