+++ src/game/scenes/main/main-menu/main-menu-controller.ts
@@ -1,5 +1,5 @@
-import { APIService } from "../../services/network/api-service.js";
-import type { MessagesResponse } from "../../interfaces/responses/messages-response.js";
+import { APIService } from "../../../services/network/api-service.js";
+import type { MessagesResponse } from "../../../interfaces/responses/messages-response.js";
 
 export class MainMenuController {
   constructor(private readonly apiService: APIService) {}
