+++ src/game/scenes/main/main-menu/main-menu-scene.ts
@@ -1,18 +1,18 @@
-import { CloseableMessageEntity } from "../../entities/common/closeable-message-entity.js";
-import { MenuOptionEntity } from "../../entities/common/menu-option-entity.js";
-import { OnlinePlayersEntity } from "../../entities/online-players-entity.js";
-import { ServerMessageWindowEntity } from "../../entities/server-message-window-entity.js";
-import { APIService } from "../../services/network/api-service.js";
-import type { MessagesResponse } from "../../interfaces/responses/messages-response.js";
-import { BaseGameScene } from "../../../core/scenes/base-game-scene.js";
-import { LoadingScene } from "../loading/loading-scene.js";
-import { ScoreboardScene } from "./scoreboard-scene.js";
-import { SettingsScene } from "./settings-scene.js";
-import { EventType } from "../../enums/event-type.js";
-import type { GameState } from "../../../core/models/game-state.js";
-import type { OnlinePlayersPayload } from "../../interfaces/events/online-players-payload.js";
-import { container } from "../../../core/services/di-container.js";
-import { EventConsumerService } from "../../../core/services/gameplay/event-consumer-service.js";
+import { CloseableMessageEntity } from "../../../entities/common/closeable-message-entity.js";
+import { MenuOptionEntity } from "../../../entities/common/menu-option-entity.js";
+import { OnlinePlayersEntity } from "../../../entities/online-players-entity.js";
+import { ServerMessageWindowEntity } from "../../../entities/server-message-window-entity.js";
+import { APIService } from "../../../services/network/api-service.js";
+import type { MessagesResponse } from "../../../interfaces/responses/messages-response.js";
+import { BaseGameScene } from "../../../../core/scenes/base-game-scene.js";
+import { LoadingScene } from "../../loading/loading-scene.js";
+import { ScoreboardScene } from "../scoreboard/scoreboard-scene.js";
+import { SettingsScene } from "../settings-scene.js";
+import { EventType } from "../../../enums/event-type.js";
+import type { GameState } from "../../../../core/models/game-state.js";
+import type { OnlinePlayersPayload } from "../../../interfaces/events/online-players-payload.js";
+import { container } from "../../../../core/services/di-container.js";
+import { EventConsumerService } from "../../../../core/services/gameplay/event-consumer-service.js";
 import { MainMenuEntityFactory } from "./main-menu-entity-factory.js";
 import type { MainMenuEntities } from "./main-menu-entity-factory.js";
 import { MainMenuController } from "./main-menu-controller.js";
