+++ src/game/scenes/main/scoreboard/scoreboard-controller.ts
@@ -1,5 +1,5 @@
-import type { RankingResponse } from "../../interfaces/responses/ranking-response.js";
-import { APIService } from "../../services/network/api-service.js";
+import type { RankingResponse } from "../../../interfaces/responses/ranking-response.js";
+import { APIService } from "../../../services/network/api-service.js";
 
 export class ScoreboardController {
   constructor(private readonly apiService: APIService) {}
