+++ src/game/scenes/main/scoreboard/scoreboard-scene.ts
@@ -1,10 +1,10 @@
-import type { RankingResponse } from "../../interfaces/responses/ranking-response.js";
-import { BaseGameScene } from "../../../core/scenes/base-game-scene.js";
-import type { GameState } from "../../../core/models/game-state.js";
-import { APIService } from "../../services/network/api-service.js";
+import type { RankingResponse } from "../../../interfaces/responses/ranking-response.js";
+import { BaseGameScene } from "../../../../core/scenes/base-game-scene.js";
+import type { GameState } from "../../../../core/models/game-state.js";
+import { APIService } from "../../../services/network/api-service.js";
 import { injectable } from "@needle-di/core";
-import { container } from "../../../core/services/di-container.js";
-import { EventConsumerService } from "../../../core/services/gameplay/event-consumer-service.js";
+import { container } from "../../../../core/services/di-container.js";
+import { EventConsumerService } from "../../../../core/services/gameplay/event-consumer-service.js";
 import { ScoreboardEntityFactory } from "./scoreboard-entity-factory.js";
 import type { ScoreboardEntities } from "./scoreboard-entity-factory.js";
 import { ScoreboardController } from "./scoreboard-controller.js";
